requirements.txt
# requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
openai>=1.12.0
azure-storage-blob>=12.19.1
azure-search-documents==11.3.0
azure-core>=1.29.5
numpy>=1.26.0
langchain>=0.1.0                # LangChain 기본 라이브러리
langchain-core>=0.1.0           # LangChain 코어 기능
langchain-community>=0.0.10     # Azure OpenAI, Azure AI Search 커넥터 포함
langchain-openai>=0.0.9         # Azure OpenAI 연결용